// src/app/generated/dcc/model/formulaDto.ts
/**
 * OpenAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FormulaDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    content?: string;
    type?: FormulaDto.TypeEnum;
}
export namespace FormulaDto {
    export type TypeEnum = 'MATHML' | 'LATEX';
    export const TypeEnum = {
        Mathml: 'MATHML' as TypeEnum,
        Latex: 'LATEX' as TypeEnum
    };
}