// src/app/generated/dcc/model/richContentDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { FormulaDto } from './formulaDto';
import { ByteDataDto } from './byteDataDto';


export interface RichContentDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    name?: LanguageSpecificStringsDto;
    textContent?: LanguageSpecificStringsDto;
    byteDataContent?: ByteDataDto;
    formulaContent?: FormulaDto;
}